+++ cucumber-expressions/javascript/test/cucumber_expression_test.js
@@ -65,8 +65,8 @@ describe(CucumberExpression.name, () => {
   describe('RegExp special characters', () => {
     ['\\', '[', ']', '^', '$', '.', '|', '?', '*', '+'].forEach((character) => {
       it(`escapes ${character}`, () => {
-        const expr = `I have {n:int} cuke(s) and ${character}`
-        const expression = new CucumberExpression(expr, [], new TransformLookup())
+        const expr = `I have {int} cuke(s) and ${character}`
+        const expression = new CucumberExpression(expr, [], new ParameterRegistry())
         const arg1 = expression.match(`I have 800 cukes and ${character}`)[0]
         assert.equal(arg1.offset, 7)
         assert.equal(arg1.value, "800")
@@ -75,7 +75,7 @@ describe(CucumberExpression.name, () => {
 
     it(`escapes .`, () => {
       const expr = `I have {n:int} cuke(s) and .`
-      const expression = new CucumberExpression(expr, [], new TransformLookup())
+      const expression = new CucumberExpression(expr, [], new ParameterRegistry())
       assert.equal(expression.match(`I have 800 cukes and 3`), null)
       const arg1 = expression.match(`I have 800 cukes and .`)[0]
       assert.equal(arg1.offset, 7)
@@ -84,7 +84,7 @@ describe(CucumberExpression.name, () => {
 
     it(`escapes |`, () => {
       const expr = `I have {n:int} cuke(s) and a|b`
-      const expression = new CucumberExpression(expr, [], new TransformLookup())
+      const expression = new CucumberExpression(expr, [], new ParameterRegistry())
       assert.equal(expression.match(`I have 800 cukes and a`), null)
       assert.equal(expression.match(`I have 800 cukes and b`), null)
       const arg1 = expression.match(`I have 800 cukes and a|b`)[0]